+++ CHANGELOG.md
@@ -1,1 +1,72 @@
+# [1.0.0-alpha.8](https://github.com/gravitee-io/gravitee-policy-js/compare/1.0.0-alpha.7...1.0.0-alpha.8) (2026-04-02)
+
+
+### Bug Fixes
+
+* restore event loop thread after schedulers.io execution ([116216a](https://github.com/gravitee-io/gravitee-policy-js/commit/116216a10670d7b2353c521a8a4f7f8f6ae2dd55))
+
+# [1.0.0-alpha.7](https://github.com/gravitee-io/gravitee-policy-js/compare/1.0.0-alpha.6...1.0.0-alpha.7) (2026-04-01)
+
+
+### Bug Fixes
+
+* **deps:** clean project configuration ([f1a6e45](https://github.com/gravitee-io/gravitee-policy-js/commit/f1a6e456ccd3a629d84eb20ce692e69c7f7400f6))
+
+# [1.0.0-alpha.6](https://github.com/gravitee-io/gravitee-policy-js/compare/1.0.0-alpha.5...1.0.0-alpha.6) (2026-03-23)
+
+
+### Features
+
+* add javascript policy icon ([216a1b2](https://github.com/gravitee-io/gravitee-policy-js/commit/216a1b2f97ae9d20ccd69f0fdc40a1e5e31c825a))
+
+# [1.0.0-alpha.5](https://github.com/gravitee-io/gravitee-policy-js/compare/1.0.0-alpha.4...1.0.0-alpha.5) (2026-03-18)
+
+
+### Bug Fixes
+
+* clean up schema-form for new console UI ([dd69cfe](https://github.com/gravitee-io/gravitee-policy-js/commit/dd69cfe80b460f61c515d92f5931a563b7d345aa))
+
+
+### Features
+
+* make console.log/console.error configurable via gravitee.yml ([db84a38](https://github.com/gravitee-io/gravitee-policy-js/commit/db84a38de963185544cc9a13bd6ed27c7c36d041))
+
+# [1.0.0-alpha.4](https://github.com/gravitee-io/gravitee-policy-js/compare/1.0.0-alpha.3...1.0.0-alpha.4) (2026-03-16)
+
+
+### Features
+
+* add configurable script timeout via gravitee.yml ([240fde3](https://github.com/gravitee-io/gravitee-policy-js/commit/240fde3da1a01cce2e1cb7b9888950e9aee32861))
+
+# [1.0.0-alpha.3](https://github.com/gravitee-io/gravitee-policy-js/compare/1.0.0-alpha.2...1.0.0-alpha.3) (2026-03-16)
+
+
+### Bug Fixes
+
+* add null guards on atob/btoa with integration tests ([bb7e80c](https://github.com/gravitee-io/gravitee-policy-js/commit/bb7e80c9cfb95b804366b259af0e3a1cabc0f61d))
+
+# [1.0.0-alpha.2](https://github.com/gravitee-io/gravitee-policy-js/compare/1.0.0-alpha.1...1.0.0-alpha.2) (2026-03-11)
+
+
+### Bug Fixes
+
+* address PR review feedback ([c499b3c](https://github.com/gravitee-io/gravitee-policy-js/commit/c499b3c02d622e1bd31d40237f7ff4cd2b7dde83))
+
+
+### Features
+
+* add JS bindings and context API ([e227e26](https://github.com/gravitee-io/gravitee-policy-js/commit/e227e263b45b66a4ff0a2f780d571b88d87bbb61))
+
+# 1.0.0-alpha.1 (2026-03-05)
+
+
+### Bug Fixes
+
+* remove stacktrace from warn log to avoid duplication ([e03b7c2](https://github.com/gravitee-io/gravitee-policy-js/commit/e03b7c250d67238a75459b2a0945f9cb9179fd80))
+
+
+### Features
+
+* implement secure GraalJS execution engine ([d80b7f2](https://github.com/gravitee-io/gravitee-policy-js/commit/d80b7f22582fc7ed5a269856e671c8a25c3fe8d1))
+
 # Changelog